Convert drawer handlers to typescript
